corrige views clientes

+++ clientes/views.py
@@ -7,26 +7,27 @@ from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
 
 
-# LOGIN DO SISTEMA
+# LOGIN
 def login_usuario(request):
+
     if request.method == 'POST':
+
         username = request.POST.get('username')
         password = request.POST.get('password')
 
         user = authenticate(request, username=username, password=password)
 
-        if user is not None:
+        if user:
             login(request, user)
 
-            if user.is_staff:
+            if user.has_perm('clientes.view_cliente'):
                 return redirect('lista_clientes')
             else:
                 return redirect('cadastrar_cliente')
 
-        else:
-            return render(request, 'clientes/login.html', {
-                'erro': 'Usuário ou senha inválidos'
-            })
+        return render(request, 'clientes/login.html', {
+            'erro': 'Usuário ou senha inválidos'
+        })
 
     return render(request, 'clientes/login.html')
 
@@ -34,11 +35,13 @@ def login_usuario(request):
 # LOGOUT
 @login_required(login_url='login')
 def logout_usuario(request):
+
     logout(request)
+
     return redirect('login')
 
 
-# LISTA DE CLIENTES (ADMIN OU USUÁRIO COM PERMISSÃO)
+# LISTA CLIENTES
 @login_required
 @permission_required('clientes.view_cliente', raise_exception=True)
 def lista_clientes(request):
@@ -58,23 +61,27 @@ def lista_clientes(request):
     })
 
 
-# CADASTRAR CLIENTE (QUALQUER USUÁRIO LOGADO)
+# CADASTRAR CLIENTE
 @login_required(login_url='login')
 def cadastrar_cliente(request):
 
     if request.method == 'POST':
+
         form = ClienteForm(request.POST)
 
         if form.is_valid():
+
             form.save()
+
             messages.success(request, "Cliente cadastrado com sucesso!")
 
             if request.user.has_perm('clientes.view_cliente'):
                 return redirect('lista_clientes')
-            else:
-                return redirect('cadastrar_cliente')
+
+            return redirect('cadastrar_cliente')
 
     else:
+
         form = ClienteForm()
 
     return render(request, 'clientes/cadastrar_cliente.html', {'form': form})
@@ -88,13 +95,16 @@ def editar_cliente(request, id):
     cliente = get_object_or_404(Cliente, id=id)
 
     if request.method == 'POST':
+
         form = ClienteForm(request.POST, instance=cliente)
 
         if form.is_valid():
             form.save()
+            messages.success(request, "Cliente atualizado com sucesso!")
             return redirect('lista_clientes')
 
     else:
+
         form = ClienteForm(instance=cliente)
 
     return render(request, 'clientes/form_cliente.html', {
@@ -110,6 +120,8 @@ def excluir_cliente(request, id):
 
     cliente = get_object_or_404(Cliente, id=id)
 
-    cliente.delete()
+    if request.method == "POST":
+        cliente.delete()
+        messages.success(request, "Cliente excluído com sucesso!")
 
     return redirect('lista_clientes')
